adição do crud de aluno

+++ back-end/src/main/java/com/example/camp_xadrez/api/domain/aluno/DadosDetalhamentoAluno.java
@@ -3,7 +3,7 @@ package com.example.camp_xadrez.api.domain.aluno;
 public record DadosDetalhamentoAluno(
         Long id,
         String nome,
-        Integer pontos_total
+        Float pontos_total
 ) {
     public DadosDetalhamentoAluno(Aluno aluno){
         this(
